refactor: split core/lib into provenance/report/actions subdirectories
