+++ core/lib/actions/annotation.js
