+++ core/lib/actions/docker-error.test.js
