+++ core/lib/provenance/errors.js
