+++ core/lib/provenance/image-ref.js
