+++ core/lib/provenance/local-image-override.js
