+++ core/lib/provenance/local-image-override.test.js
