+++ core/lib/provenance/verify-image.js
