+++ core/lib/report/build-example.js
