+++ core/lib/report/build-example.test.js
