+++ core/lib/report/host-table.js
