+++ core/lib/report/host-table.test.js
