+++ core/lib/report/markdown-table.js
