+++ docs/security.md
@@ -448,4 +448,4 @@ The binding of the image digest to the exact source commit SHA also serves as an
 - **Account compromise**: If the repository owner's GitHub account or the repository itself is compromised, an attacker could trigger the release workflow and produce a legitimately-signed malicious image.
 - **Trust in Sigstore infrastructure**: Verification relies on the availability and integrity of the Rekor transparency log and the Fulcio certificate authority. The TUF-backed trust root is fetched at verification time; a network outage will cause the main phase to hard-fail.
 - **TOCTOU window**: The manifest digest is fetched before the bundle is pulled. A highly targeted attack that replaces the registry content in the window between these two steps would still succeed — though such an attack requires compromising the registry itself. Note that the subsequent `docker pull` is digest-pinned (`image@sha256:…`), so there is no TOCTOU between verification and the actual image pull; the residual window is limited to between the manifest-digest fetch and the bundle fetch.
-- **Local/CI self-test bypass**: `BUILDCAGE_BUILD_TEST_HOOKS=1 pnpm build` compiles a `setup/dist/main.cjs` where a `BUILDCAGE_LOCAL_IMAGE_REF` escape hatch is reachable, used only by this repo's own `test_action` CI job and local development. The bypass logic lives in its own module (`core/lib/local-image-override.js`), loaded only via a dynamic `import()` behind this build-time flag; in every normal build (including every published release), rollup's own module-graph tree-shaking excludes that entire file from the bundle — not just the call to it. A consumer of the published `dash14/buildcage/setup@<ref>` action cannot reach it no matter what `env:` they set, since the code is physically absent from what they execute. `unit_test`'s CI job additionally asserts, by inspecting the built file directly, that a normal build never contains a live runtime read of `BUILDCAGE_BUILD_TEST_HOOKS` — guarding against a future refactor silently breaking that guarantee. See [development.md](./development.md#local-development) for details.
+- **Local/CI self-test bypass**: `BUILDCAGE_BUILD_TEST_HOOKS=1 pnpm build` compiles a `setup/dist/main.cjs` where a `BUILDCAGE_LOCAL_IMAGE_REF` escape hatch is reachable, used only by this repo's own `test_action` CI job and local development. The bypass logic lives in its own module (`core/lib/provenance/local-image-override.js`), loaded only via a dynamic `import()` behind this build-time flag; in every normal build (including every published release), rollup's own module-graph tree-shaking excludes that entire file from the bundle — not just the call to it. A consumer of the published `dash14/buildcage/setup@<ref>` action cannot reach it no matter what `env:` they set, since the code is physically absent from what they execute. `unit_test`'s CI job additionally asserts, by inspecting the built file directly, that a normal build never contains a live runtime read of `BUILDCAGE_BUILD_TEST_HOOKS` — guarding against a future refactor silently breaking that guarantee. See [development.md](./development.md#local-development) for details.
