+++ run/src/main.js
@@ -4,10 +4,10 @@ import { join, dirname } from "node:path";
 import { fileURLToPath } from "node:url";
 
 import { parseAndValidateRules } from "../../core/shared/lib/rules.js";
-import { resolveBuildcageImageRef } from "../../core/lib/image-ref.js";
-import { verifyImageDigest } from "../../core/lib/verify-image.js";
-import { describeDockerFailure } from "../../core/lib/docker-error.js";
-import { createAnnotation } from "../../core/lib/annotation.js";
+import { resolveBuildcageImageRef } from "../../core/lib/provenance/image-ref.js";
+import { verifyImageDigest } from "../../core/lib/provenance/verify-image.js";
+import { describeDockerFailure } from "../../core/lib/actions/docker-error.js";
+import { createAnnotation } from "../../core/lib/actions/annotation.js";
 import { SandboxError } from "./lib/errors.js";
 import { checkPasswordlessSudo } from "./lib/sudo-preflight.js";
 import { generateContainerName, getContainerPid, deriveProjectName } from "./lib/container.js";
@@ -37,7 +37,7 @@ const LOCAL_IMAGE_OVERRIDE_ENABLED = process.env.BUILDCAGE_BUILD_TEST_HOOKS === 
 /**
  * Verifies image provenance and resolves the digest-pinned image ref for
  * the run action's (buildkitd-less) proxy image, published under the `-proxy`
- * tag suffix (see imageTagFromRef in core/lib/verify-image.js).
+ * tag suffix (see imageTagFromRef in core/lib/provenance/verify-image.js).
  */
 async function resolveVerifiedImage({ actionRef, actionRepo }) {
   let digest;
@@ -85,7 +85,7 @@ export function buildACLRules({ httpsRulesInput, httpRulesInput, ipRulesInput })
 }
 
 /**
- * Never sent to the container's ACL — see core/lib/known-blocked.js.
+ * Never sent to the container's ACL — see core/lib/report/known-blocked.js.
  */
 export function readKnownBlockedRules(input) {
   return parseRulesOrThrow(input);
@@ -131,7 +131,7 @@ async function main() {
   const annotation = createAnnotation(Boolean(env.GITHUB_STEP_SUMMARY));
 
   const localOverride = LOCAL_IMAGE_OVERRIDE_ENABLED
-    ? (await import("../../core/lib/local-image-override.js")).readLocalImageOverride(env)
+    ? (await import("../../core/lib/provenance/local-image-override.js")).readLocalImageOverride(env)
     : null;
   if (localOverride) {
     console.log(
