+++ setup/src/main.js
@@ -4,9 +4,9 @@ import { fileURLToPath } from "node:url";
 
 import { buildRules } from "../../core/shared/lib/rules.js";
 import { SetupError } from "./lib/errors.js";
-import { verifyImageDigest } from "../../core/lib/verify-image.js";
-import { resolveBuildcageImageRef } from "../../core/lib/image-ref.js";
-import { describeDockerFailure } from "../../core/lib/docker-error.js";
+import { verifyImageDigest } from "../../core/lib/provenance/verify-image.js";
+import { resolveBuildcageImageRef } from "../../core/lib/provenance/image-ref.js";
+import { describeDockerFailure } from "../../core/lib/actions/docker-error.js";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const composeFile = join(__dirname, "../compose.yaml");
@@ -14,7 +14,7 @@ const composeFile = join(__dirname, "../compose.yaml");
 // Gates a local-image override used only by this repo's own CI/dev testing
 // (see test_action in .github/workflows/test-e2e.yml), never by a consumer of
 // a published action. A normal build physically excludes
-// core/lib/local-image-override.js (rollup tree-shakes the dead import); the
+// core/lib/provenance/local-image-override.js (rollup tree-shakes the dead import); the
 // unit_test CI job also greps the built output as a backstop.
 const LOCAL_IMAGE_OVERRIDE_ENABLED = process.env.BUILDCAGE_BUILD_TEST_HOOKS === "1";
 
@@ -53,7 +53,7 @@ async function main() {
   console.log(`Proxy engine: ${proxyEngine}`);
 
   const localOverride = LOCAL_IMAGE_OVERRIDE_ENABLED
-    ? (await import("../../core/lib/local-image-override.js")).readLocalImageOverride(env)
+    ? (await import("../../core/lib/provenance/local-image-override.js")).readLocalImageOverride(env)
     : null;
   if (localOverride) {
     console.log(
