+++ setup/src/main.property.test.js
@@ -7,8 +7,8 @@ import { describe, it } from "node:test";
 import assert from "node:assert/strict";
 import fc from "fast-check";
 
-import { imageTagFromRef } from "../../core/lib/verify-image.js";
-import { resolveBuildcageImageRef } from "../../core/lib/image-ref.js";
+import { imageTagFromRef } from "../../core/lib/provenance/verify-image.js";
+import { resolveBuildcageImageRef } from "../../core/lib/provenance/image-ref.js";
 import { buildACLRules, resolveProxyEngine } from "./main.js";
 
 // ---------------------------------------------------------------------------
